backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.1
pika==1.3.2
Werkzeug==3.0.1
python-engineio==4.7.1
python-socketio==5.9.0
